Stop hardcoding string indices in the HID report descriptor
Done, to make it easier to support device-specific string reports when exposing multiple HID devices in a USB composite device.

The string indices in the HID report descriptor doesn't seem to be used for anything, since the actual string indices are retrieved through a HID FEATURE report report. The change therefore doesn't seem to break anything.