Defer UpdateBatteryState call to a read request completion handler
Fixes #79 by allowing the HW to first fill in the request _before_ starting to parse the associated buffer.

The weird thing is that the current implementation that calls ParseReadHidBuffer directly from the read request handler seemed to work fine. That's probably because it's reading the previous INPUT report as suggested in https://community.osr.com/t/how-to-alter-a-hid-report-in-a-filter-driver/41232/2.